gcc warnings
Former-commit-id: 9089de8d251c6ba16af86f4c44a50d30e3b0be1c

+++ iwave/base/src/parser/convert.c
@@ -22,7 +22,7 @@ static inline int flt_start(char *str)
     if ( (*str == '+') || (*str == '-') ) ++str;
     /* TODO: check inf here */
     if ( *str == '.' ) ++str;
-    return isdigit(*str);
+    return isdigit((int) *str);
 }
 /*----------------------------------------------------------------------------*/
 /*
@@ -35,7 +35,7 @@ char *str :  null-terminated string.
 static inline int int_start(char *str)
 {
     if ( (*str == '+') || (*str == '-') ) ++str;
-    return isdigit(*str);
+    return isdigit((int) *str);
 }
 /*----------------------------------------------------------------------------*/
